Lotsa(tm) Changes
This index.php will now check to see if the image directory exists and if the image has not been downloaded - if both are true the images will be downloaded to the image directory. If the image does not exist the image is hotlinked from the discogs server. I will soon add a note saying that the image is being hotlinked to discourage this practice.

I will soon change the included python script to just download all cover images in the collection and get ride of the options to locally store the data files as I dont think its needed anymore.

I've cleaned up some of the code and  added some more icons to make it more visually interesting.