Tidying some more code
Made the extended information into collapsible tables on the single release view.

Starting to move logic code from the HTML to the start of the display functions, just to clean things up a bit and improve readability. Using variables in HTML.

Added "Companies" table - the last of the release data I'm going to include.

Consolidated some of the table columns - not everything needs three columns and it was hard to read on larger screens.