Added Random Album button
Thank you laminateddenim for the idea! 

I was able to figure out a way to randomize based on the entirety of the collection (or whichever folder you're currently viewing) instead of just the current paginated set of items.

It works but I'm pretty certain I duplicated some of the code while creating the get_random_release function. It's late and I need to go to bed so I'm going to look into cleaning it up and refactoring some stuff this week.