Update index.php
Lots of changes - but still a work in progress.

New:
Got rid of cached files, except for images, and a copy of the entire collection as JSON which will be used for dynamic search.

Added dynamic search. Currently doesn't get updated automatically - I need to figure out how I want to improve this.

Cleaned up a lot of the HTML.

Todo: 
Clean up unused variables. 
Figure out a better way to download cached images - currently it downloads them on the fly but frequently hits API limits if you're missing lots of them.